Define UNDEF object for use in rspec files
Despite what you may read online, setting a param to `nil` in
Ruby is NOT the same thing as setting it to `undef` in Puppet.
The only good solution to this that I've found was proposed at
the end of this thread [1].

Creating a special class / instance in spec_helper.rb can help
work around this issue.

[1] <https://groups.google.com/forum/#!topic/puppet-users/6nL2eROH8is>