Set $KCODE='UTF-8' in spec/spec_helper.rb
This will allow us to write tests that pass unicode string
parameters to puppet.

This configuration tells Ruby to use UTF-8 encoding to parse
Ruby .rb files, and to not escape non-ASCII characters in
unicode strings when printing and calling `#inspect`.

Ruby 1.x uses ASCII, not UTF-8, by default. When non-ASCII
characters are printed in the REPL, they are not printed
natively, but instead the escaped code units are printed. This
means that `#inspect` returns an ASCII string where the
backslashes in the code units are themselves escaped.

As an example, the string 'ƒőő' is printed in the REPL as
\306\222\305\221\305\221. So 'ƒőő'.inspect returns
'\'\\306\\222\\305\\221\\305\\221\''. With $KCODE='UTF-8', 'ƒőő'
is printed in the REPL as ƒőő, and 'ƒőő'.inspect returns
'\'ƒőő\''.

When using rspec-puppet to pass `value` as a parameter from a
spec file to puppet, the value that is actually passed is
`"#{value.inspect}"` [1]. When testing a non-ASCII unicode
string parameter, something like
'\\306\\222\\305\\221\\305\\221' is not actually the string
that was intended to be passed to puppet; 'ƒőő' is the intended
value to use.

References:
[1] <https://github.com/rodjek/rspec-puppet/blob/v2.1.0/lib/rspec-puppet/support.rb#L119>
[2] <https://www.ruby-forum.com/topic/144111>
[3] <https://www.ruby-forum.com/topic/178589>