Define contexts for Windows, POSIX tests
This allows developers to write tests for functionality that
depends on the agent's platform (perhaps because it depends on a
fact, or because it declares a resource with an absolute path
parameter), regardless of the test machine's platform.

In order to make this possible, it is necessary to extend
RSpec::Core::MemoizedHelpers, in order to define a version of
`let` that can be updated multiple times within the same nested
context level.