Fix Windows resources in unit tests
Fix loading of Windows resource providers and puppetforge
resource types when faking Windows for unit tests.

The problem was that if the first test was a fake Windows test,
then during catalog compilation, all of the resource types would
get loaded for the first time, pass through OS checks, and fail
when trying to import Windows libraries that don't exist on
Linux.

The solution is to force loading of these types before the test
runs and even before the OS is faked, by doing a few
compilations of the catalog.

There was also a problem with the windows provider of the file
resource type. There would be problems for Windows tests
regardless of when it was loaded. The hacky solution is to
custom define it during tests; I couldn't find a better
solution.