In tests, work around Acl's Ruby 1.9+ req
Patch RSpec::Puppet::ManifestMatchers::Compile to not fail when
encountering certain exceptions raised by the Acl resource type
on Ruby 1.8.

The puppetlabs/acl module [1] provides the Acl resource type,
which is needed by manifests that need to set file permissions
on Windows. Although it is not documented, this module requires
Ruby 1.9+.

It turns out that these Ruby 1.8-incompatible code paths are
only executed on the Puppet client. All of the puppetlabs/acl
code that is executed on the Puppet master is Ruby 1.8
compliant.

However, the rspec-puppet test ``it { should compile }``
executes this client-side code, and is negatively affected by
the incompatibility, if the Puppet test machines are running
Ruby 1.8. This would cause the compilation test to fail for
manifests that would successfully compile in production.

Rather than dissuade use of ``it { should compile }``, patch the
compile test when tests are being run on Ruby 1.8. If any of the
known exceptions are raised during a compilation test, the test
is skipped, with an explanation displayed to the user. And when
systems upgrade to Ruby 1.9, the tests will begin to work again.

[1] <https://forge.puppetlabs.com/puppetlabs/acl>