Patch Support#setup_puppet()
Up through version 2.2.0, rspec-puppet's implementation of
Support#setup_puppet() could not correctly handle multiple
module path directories [1]. As of the time of this writing,
this bug has been fixed in the master branch, but there has not
yet been a release containing this fix. Override that method
with the implementation from the master branch. And move the
previous patch for Support#import_str() into the same file as
this patch.

[1] https://github.com/rodjek/rspec-puppet/pull/223