Add let_params_merge() method
This method is analogous to the existing `let_facts_merge()`
method, except that it is used to define class/define params for
a test.

This is a developer convenience method. It is easier to use than
``let_hash_merge(:params, hash)``, and better than
``let(:params) { hash }`` (see the documentation in
update_let_definition_support.rb for details).