Make test failures fail faster
Before doing platform stubbing, the catalog is compiled a few
times first, to exercise any Ruby definitions that need to be
executed before doing the stubbing.

If the same exception is seen twice in a row, compiling
additional times is going to have the same result. So we fail
faster by breaking out of the loop early.