Enable hiera lookups during unit tests
Unit tests were failing because hiera wasn't enabled during unit
testing.

Manually created a hiera.yaml file in spec/fixtures that mirrors
what the real hiera.yaml file on the puppet masters would look
like. As a TODO for the future, when we have a real cluster of
test machines, we can put the real hiera.yaml file on those
machines and use that instead of a manually-maintained file.

This change shouldn't be necessary, because component modules
shouldn't be using hiera at all. They should instead accept
parameters which are passed from profiles, which may look up
values in hiera. But given that our modules aren't currently
independent, and that they do use hiera, this change is
necessary at this time.

The ideal long term fix is to refactor all component modules to
not call hiera directly and to instead parameterize all required
data. At that point, the hiera.yaml file and the hiera_config
setting can be removed to help expose offending modules.

Also fix some tests that were failing with

    The  OS on foo is not supported by puppet

because they were stubbing out the operatingsystem fact, but not
the osfamily or kernel facts.